Move form to layout
